refactor: ship the unified scanner inline, gated by parity
A repo-root lib/ never ships with a single-skill install, so the
runtime import crashed installed relays (test/relay-isolated.mjs: 8
failures). The unified scanner now ships inline in each relay that
uses it — MAX_BUFFERED_CHARS in all 8, makeEventScanner in 7,
makeLineScanner in vibe — as byte-identical copies gated by
relay-parity, which now supports per-symbol relay subsets. The scanner
implementation itself is unchanged from review. Its unit tests run
against the source extracted from claude's/vibe's inlined copies, and
CI runs them before smoke. Isolated install: 10/10.
